+++ qase-java-commons/src/main/java/io/qase/commons/models/report/ReportResult.java
@@ -23,6 +23,7 @@ public class ReportResult {
     
     public ReportExecution execution;
     public Map<String, String> fields;
+    public List<String> tags;
     public List<ReportAttachment> attachments;
     public List<ReportStepResult> steps;
     public Map<String, String> params;